api/requirements.txt
aniso8601==7.0.0
click==7.1.2
dnspython==2.1.0
email-validator==1.1.2
Flask==1.1.2
Flask-GraphQL==2.0.1
Flask-JWT-Extended==3.25.0
flask-mongoengine==1.0.0
Flask-WTF==0.14.3
graphene==2.1.8
graphene-mongo==0.2.13
graphql-core==2.3.2
graphql-relay==2.0.1
graphql-server-core==1.2.0
idna==3.1
iso8601==0.1.13
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mongoengine==0.22.1
passlib==1.7.4
promise==2.3
PyJWT==1.7.1
pymongo==3.11.2
python-dotenv==0.15.0
Rx==1.6.1
singledispatch==3.4.0.3
six==1.15.0
Werkzeug==1.0.1
WTForms==2.3.3